scripts/benchmark.py
#!/usr/bin/env python3
"""
Main benchmarking script for Llama 3 LoRA configurations.
"""

import argparse
import yaml
import json
import time
from pathlib import Path
from datetime import datetime
import torch
from transformers import (
    AutoModelForCausalLM,
    AutoTokenizer,
    TrainingArguments,
    Trainer,
    DataCollatorForLanguageModeling
)
from peft import LoraConfig, get_peft_model, TaskType
from datasets import load_dataset
import wandb

def parse_args():
    parser = argparse.ArgumentParser()
    parser.add_argument("--config", type=str, required=True)
    parser.add_argument("--model_name", type=str, default="meta-llama/Llama-3.2-1B")
    parser.add_argument("--output_dir", type=str, default="results/raw")
    parser.add_argument("--use_wandb", action="store_true")
    return parser.parse_args()

def load_config(config_path):
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)

def prepare_model_and_tokenizer(model_name, lora_config):
    """Load model with LoRA configuration."""
    tokenizer = AutoTokenizer.from_pretrained(model_name)
    tokenizer.pad_token = tokenizer.eos_token
    
    model = AutoModelForCausalLM.from_pretrained(
        model_name,
        torch_dtype=torch.float16,
        device_map="auto",
    )
    
    # Apply LoRA
    peft_config = LoraConfig(
        task_type=TaskType.CAUSAL_LM,
        **lora_config
    )
    model = get_peft_model(model, peft_config)
    model.print_trainable_parameters()
    
    return model, tokenizer

def benchmark_configuration(config, model_name, output_dir):
    """Run benchmark for a single configuration."""
    print(f"Benchmarking: {config}")
    
    # Track metrics
    metrics = {
        "config": config,
        "model": model_name,
        "start_time": datetime.now().isoformat(),
    }
    
    # Load model
    start_time = time.time()
    model, tokenizer = prepare_model_and_tokenizer(model_name, config["lora"])
    metrics["model_load_time"] = time.time() - start_time
    
    # Memory tracking
    if torch.cuda.is_available():
        torch.cuda.reset_peak_memory_stats()
        metrics["gpu_available"] = torch.cuda.get_device_name()
    
    # TODO: Add training loop
    # TODO: Add evaluation
    # TODO: Save results
    
    return metrics

def main():
    args = parse_args()
    config = load_config(args.config)
    
    # Run benchmarks
    results = benchmark_configuration(
        config,
        args.model_name,
        args.output_dir
    )
    
    # Save results
    output_path = Path(args.output_dir) / f"benchmark_{datetime.now().strftime('%Y%m%d_%H%M%S')}.json"
    output_path.parent.mkdir(parents=True, exist_ok=True)
    
    with open(output_path, 'w') as f:
        json.dump(results, f, indent=2)
    
    print(f"Results saved to {output_path}")

if __name__ == "__main__":
    main()
